refactor(sequel): extract optimized SQL helpers
Issue: No ticket/issue
